requirements.txt
beautifulsoup4~=4.13
requests~=2.32
pyyaml~=6.0